requirements.txt
python-whois==0.9.5
httpx[asyncio]==0.27.0
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
streamlit>=1.32.0
markdown>=3.5.0
weasyprint>=62.0
